// mint-app/src/utils/enums/project/project-list.enum.ts
export enum PROJECT_STATUS {
  REQUEST_COMPLETED = 'REQUEST_COMPLETED',
  EXPERT_MATCHING_IN_PROGRESS = 'EXPERT_MATCHING_IN_PROGRESS',
  MATCHING_COMPLETED = 'MATCHING_COMPLETED',
  UNDER_CONTRACT = 'UNDER_CONTRACT',
  PROJECT_IN_PROGRESS = 'PROJECT_IN_PROGRESS',
  PROJECT_COMPLETED = 'PROJECT_COMPLETED',
}
export enum PROJECT_STATICS_ENUM {
  ENTIRE = 'ENTIRE',
  REQUEST_COMPLETED = 'REQUEST_COMPLETED',
  EXPERT_MATCHING_IN_PROGRESS = 'EXPERT_MATCHING_IN_PROGRESS',
  MATCHING_COMPLETED = 'MATCHING_COMPLETED',
  UNDER_CONTRACT = 'UNDER_CONTRACT',
  PROJECT_IN_PROGRESS = 'PROJECT_IN_PROGRESS',
  PROJECT_COMPLETED = 'PROJECT_COMPLETED',
}